Renamed unit_tests package to tests, cleaned etcd_test.go based off review
